+++ weather-app/src/icons/foggy.js
