+++ weather-app/src/icons/rainy.js
