+++ weather-app/src/icons/snowy.js
