+++ weather-app/src/icons/stormy.js
