+++ weather-app/src/icons/sunny.js
